% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_LEEF_1.R
\name{classify_LEEF_1}
\alias{classify_LEEF_1}
\title{Classify \code{algae_traits} nad calculates densities}
\usage{
classify_LEEF_1(
  algae_traits,
  classifiers_constant,
  classifiers_increasing,
  composition,
  exp_design,
  species_tracked,
  timestamp
)
}
\arguments{
\item{algae_traits}{algae_traits}

\item{classifiers_constant}{constant temperature classifier}

\item{classifiers_increasing}{increasing temperature classifier}

\item{composition}{composition}

\item{exp_design}{experimental design}

\item{species_tracked}{species tracked}

\item{timestamp}{timestamp to be used to stamp the classified data}
}
\value{
\code{list} containing two objects:
\itemize{
\item \code{algae_traits} including species
\item \code{algae_densities} densities of the differenc particles identifieds
}
}
\description{
Classify \code{algae_traits} nad calculates densities
}
