% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcam.R
\name{extractor_flowcam}
\alias{extractor_flowcam}
\title{Preprocessor flowcam data}
\usage{
extractor_flowcam(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
extract data from all \code{*_classes_*_data.csv} files
}
