% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_classifiers}
\alias{par_classifiers}
\title{Name of classifier file for constant temperature treatment}
\usage{
par_classifiers(value)
}
\description{
Name of classifier file for constant temperature treatment
}
