% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_flowcam.R
\name{pre_processor_flowcam}
\alias{pre_processor_flowcam}
\title{Preprocessor flowcam data}
\usage{
pre_processor_flowcam(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Convert all \code{.tif} files in \code{flowcam} folder to zip compressed TIFF.
}
